Struct interledger_store_redis::RedisStore

pub struct RedisStore { /* fields omitted */ }

A Store that uses Redis as its underlying database.

This store leverages atomic Redis transactions to do operations such as balance updates.

Currently the RedisStore polls the database for the routing table and rate updates, but future versions of it will use PubSub to subscribe to updates.

Methods

impl RedisStore

pub fn get_all_accounts_ids(
&self
) -> impl Future<Item = Vec<AccountId>, Error = ()>